Refactor the main solve function
Changed how it works such that you give the solver your objective and
it does some checks before the run. Also moved the start point
requirment out of the objective protocol, which I think makes more
sense. Instead you give the solver the initial primal and dual when you
actually want to solve it.

However, I did leave the startPoint() function in the symbolic objective
as that can just be a utility to get one if you don't know a feasible
point.